requirements.txt
streamlit
python-dotenv
google-genai
requests
Pillow